+++ changelog.md
@@ -1,6 +1,7 @@
 # 1.1.0
 - Added
   - Mods tab: May now group by Enabled (default), Mod Type, Game Version, and Author.
+  - Support for MediaFire download links (thanks, AtlanticAccent & ChatGPT).
 - Fixed
   - Searching on the Mods tab also filtered the Dashboard mods.
   - Search included some less relevant results and didn't include perfect matches (e.g. searching "iron shell" didn't show it, but "iron shel" did).